Add create character form
