+++ src/Equipment/type.ts
@@ -1,5 +1,3 @@
-import type { Equipment } from "../type";
-
 export type FormData = {
   type:
     | "equipment"
