+++ src/components/Form/Fieldset.svelte
@@ -1,0 +1,10 @@
+<script type="ts">
+  export let legend: string = "";
+</script>
+
+<fieldset>
+  {#if !!legend}
+    <legend>{legend}</legend>
+  {/if}
+  <slot />
+</fieldset>
